# Better P256.pointMul performance