Existing implementation does unnecessary work to multiply the basepoint with a scalar which is always zero. Using a specialized function skipping this saves 20-25% of the total time.

Before:
```
P256/pointAddTwoMuls-P256                mean 915.2 μs  ( +- 2.139 μs  )
P256/pointAdd-P256                       mean 28.51 μs  ( +- 92.76 ns  )
P256/pointMul-P256                       mean 443.5 μs  ( +- 1.059 μs  )
```

After:
```
P256/pointAddTwoMuls-P256                mean 719.6 μs  ( +- 1.809 μs  )
P256/pointAdd-P256                       mean 28.47 μs  ( +- 82.05 ns  )
P256/pointMul-P256                       mean 345.1 μs  ( +- 670.0 ns  )
```